+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"_info":"// This is a generated file; do not edit or check into version control.","dependencyGraph":[{"name":"flutter_current_locale","dependencies":[]}]}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_current_locale","path":"/home/skela/code/olavstoppen/emsp/plugins/flutter_current_locale/","dependencies":[]}],"android":[{"name":"flutter_current_locale","path":"/home/skela/code/olavstoppen/emsp/plugins/flutter_current_locale/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"flutter_current_locale","dependencies":[]}],"date_created":"2021-05-11 10:35:46.505094","version":"2.0.6"}
