+++ lib/current_locale.dart
@@ -1,188 +1,28 @@
-import 'dart:async';
-import 'dart:io';
+
+import 'current_locale_platform.dart' if (dart.library.js) "current_locale_web.dart";
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
-import 'dart:html' as html;
+
 
 class CurrentLocale
 {
-  static const String kChannelName = "plugins.olavstoppen.no/current_locale";
+  final CurrentLocaleManager _manager = CurrentLocaleFactory.createManager();
 
-  bool get isPlatformSupported
-  {
-    if(kIsWeb) return true;
-    return Platform.isIOS || Platform.isAndroid;
-  }
+  bool get isPlatformSupported => _manager.isPlatformSupported;
 
-  Future<String> getCurrentLanguage() async
-  {
-    if(kIsWeb)
-    {
-      var language = html.window.navigator.language;
-      if(language == null) return "nb";
-      switch(language)
-      {
-        case "nb-NO":
-        case "nb":
-        case "no":
-        case "nn": return "nb";
-        case "en-US":
-        case "en": return "en";
-      }
-      return "nb";
-    }
+  Future<String> getCurrentLanguage() => _manager.getCurrentLanguage();
 
-    return _invokeMethod("getCurrentLanguage");
-  }
+  Future<String> getCurrentCountryCode() => _manager.getCurrentCountryCode();
 
-  Future<String> getCurrentCountryCode() async
-  {
-    if(kIsWeb)
-    {
-      var language = html.window.navigator.language;
-      if(language == null) return "NO";
-      switch(language)
-      {
-        case "nb-NO":
-        case "nb":
-        case "no":
-        case "nn": return "NO";
-        case "en-US":
-        case "en": return "US";
-      }
-      return "NO";
-    }
+  Future<CurrentLocaleResult> getCurrentLocale() => _manager.getCurrentLocale();
+}
 
-    return _invokeMethod("getCurrentCountryCode");
-  }
-
-  CurrentLocaleResult get noCurrentLocalResult
-  {
-    return CurrentLocaleResult(
-        identifier: "nb_NO",
-        decimals: ",",
-        language: CurrentLocaleInfo(phone: "nb",locale: "no"),
-        country: CurrentCountryInfo(phone: null, locale: "NO", region: "NO")
-    );
-  }
-
-  CurrentLocaleResult get enCurrentLocaleResult
-  {
-    return CurrentLocaleResult(
-        identifier: "en_US",
-        decimals: ",",
-        language: CurrentLocaleInfo(phone: "en",locale: "en"),
-        country: CurrentCountryInfo(phone: null, locale: "US", region: "US")
-    );
-  }
-
-  Future<CurrentLocaleResult> getCurrentLocale() async
-  {
-    if(kIsWeb)
-    {
-      var language = html.window.navigator.language;
-      if(language == null) return Future.value(noCurrentLocalResult);
-      switch(language)
-      {
-        case "nb-NO":
-        case "nb":
-        case "no":
-        case "nn": return Future.value(noCurrentLocalResult);
-        case "en-US":
-        case "en": return Future.value(enCurrentLocaleResult);
-      }
-      return Future.value(noCurrentLocalResult);
-    }
-
-
-    var supported = isPlatformSupported;
-    if (!supported)
-      return Future.value(null);
-
-    var platform = MethodChannel(kChannelName);
-    try
-    {
-      dynamic result = await platform.invokeMethod("getCurrentLocale");
-      if (result != null && result is Map)
-      {
-        String getString(String key)
-        {
-          String identifier;
-          if (result.containsKey(key))
-            identifier = result[key];
-          return identifier;
-        }
-
-        CurrentLocaleInfo getLocale(String key)
-        {
-          final fallback = CurrentLocaleInfo();
-          if (!result.containsKey(key)) return fallback;
-          try{
-            var d = Map<String, String>.from(result[key]);
-            String phone;
-            String locale;
-            if (d.containsKey("phone"))
-              phone = d["phone"];
-            if (d.containsKey("locale"))
-              locale = d["locale"];
-            return CurrentLocaleInfo(phone: phone, locale: locale);
-          } catch (e)
-          {
-            return fallback;
-          }
-        }
-        CurrentCountryInfo getCountry(String key)
-        {
-          var info = getLocale(key);
-          try
-          {
-            var d = Map<String, String>.from(result[key]);
-            String region;
-            if (d.containsKey("region"))
-              region = d["region"];
-            return CurrentCountryInfo(locale:info.locale,phone:info.phone,region:region);
-          }
-          catch (e)
-          {
-            return CurrentCountryInfo(locale:info.locale,phone:info.phone);
-          }
-        }
-        var identifier = getString("identifier");
-        var decimals = getString("decimals");
-        var country = getCountry("country");
-        var language = getLocale("language");
-        return Future.value(CurrentLocaleResult(identifier:identifier,decimals:decimals,language:language,country:country));
-      }
-    }
-    catch (e)
-    {
-
-    }
-    return Future.value(null);
-  }
-
-  Future<String> _invokeMethod(String method) async
-  {
-    var supported = isPlatformSupported;
-    if (!supported)
-      return Future.value(null);
-
-    var platform = MethodChannel(kChannelName);
-    try
-    {
-      String result = await platform.invokeMethod(method);
-      if (result != null)
-      {
-        return Future.value(result);
-      }
-    }
-    catch (e)
-    {
-
-    }
-    return Future.value(null);
-  }
+abstract class CurrentLocaleManager
+{ 
+  bool get isPlatformSupported; 
+  Future<String> getCurrentCountryCode();
+  Future<String> getCurrentLanguage();
+  Future<CurrentLocaleResult> getCurrentLocale();
 }
 
 class CurrentLocaleResult
