+++ test/flutter_current_locale_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_current_locale/current_locale.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('flutter_current_locale');
